Wide logical IDs: every owned non-function pre-seal carrier is a checked private-u32 newtype and the wire narrows only through the measure core's policy-clean path
IMGTABLE01 cutover 1 (design of record: drafts 2+3+4, A4.3). StrId, ConstId,
TypeId, EnumId, CollTypeId widen from u16 to private u32 ordinals with checked
issuance (wide_ordinal); the durable root reference becomes the typed wide
RootId (AdmittedRoot::root_id, ImageType::Identity, MakeIdentity, LTy/compiler
DurableRoot); embedded type/instruction references (ConstLoad, Unreachable,
Todo, RecordNew, EnumConstruct.enum_idx, ListNew/MapNew/TextSplit/TextLines,
DurIterateBounded/DurIndexScan.list_ty) carry the typed ids; the Product/DAG
row ordinals (DeclarationNodeOrdinal, RootOccurrenceOrdinal, DeclarationSpan,
DeclarationMemberDef.parent) widen to u32; the coupled compiler registries
(GArg::Collection -> CollTypeId, TypeInstKey, NominalId, collection_index,
instantiate_list/map, LTy) retain the typed wide carriers.

The only narrowing direction is the measure core's policy-clean checked path:
wire_ordinal/wire_len in measure.rs, spelled checked on the fitting arm under
the existing MAX_* <= u16::MAX const-asserts in bounds; the encoder's direct
len()-as-u16 spellings for owned tables and every owned-id emission route
through it. Wire-side u16 reads widen through const from_index constructors;
the verifier/VM/runner sealed domain keeps its decoded u16 representation and
widens at the shared-type boundary. Function-slot ids (FuncId, call/export/
test halves, counts, spans, locals) and site operands are untouched
(IMGFUNC01's and cutover 3's); the within-occurrence managed-index ordinal
stays u16 as the subordinate in-row position the frozen F-1 pin's refusal arm
stands on.

Re-derived with this change and cited here: DURABLE_MEMBER_ROW_BYTES 56 -> 64
(DeclarationNode row widened) and the compiler-side max-live figure
42,850,312 -> 43,243,432 B (resource_limits pin + implementation map moved
together). instantiation_state_tests compresses to stay inside its inherited
tidy budget. Frozen tripwires byte-identical: the four full-image digests and
the 241-byte unit image, DIAMOND, counted==emitted, canonical_order, the 96
measure_verdicts pins, 35 legacy_ok_pins, 19 mutation_state pins. Full
workspace battery 3458/0; fmt, clippy -D warnings, zero unsafe, Cargo.lock
unchanged.

+++ crates/marrow-image/src/durable_id.rs
@@ -1076,7 +1076,7 @@ mod tests {
 
     /// One flat field command.
     fn field_cmd(
-        parent: Option<u16>,
+        parent: Option<u32>,
         byte: u8,
         required: bool,
         value: ValueShapeNodeId,
@@ -1091,7 +1091,7 @@ mod tests {
         }
     }
 
-    fn group_cmd(parent: Option<u16>, byte: u8) -> DeclarationMemberDef {
+    fn group_cmd(parent: Option<u32>, byte: u8) -> DeclarationMemberDef {
         DeclarationMemberDef {
             parent,
             shape: DeclarationMemberShape::Group { id: id(byte) },
@@ -1100,7 +1100,7 @@ mod tests {
 
     fn branch_cmd(
         draft: &mut ImageDraft,
-        parent: Option<u16>,
+        parent: Option<u32>,
         byte: u8,
         keys: Vec<KeyColumn>,
     ) -> DeclarationMemberDef {
